requirements.txt
# Core + Benchmark dependencies (no llama-cpp-python)

fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-dotenv>=1.0.1

numpy==1.26.4
scipy==1.11.4
scikit-learn==1.5.2
faiss-cpu==1.8.0.post1

transformers>=4.51.0
sentence-transformers>=3.3.0
accelerate==1.0.1
einops==0.8.0
huggingface-hub>=0.34.0

python-multipart==0.0.12
orjson==3.10.7
httpx==0.27.2
jinja2==3.1.6

requests>=2.31
psutil>=5.9
tqdm>=4.66

# ==========================================================
# Graph RAG
# ==========================================================
lightrag-hku>=1.0.0

# ==========================================================
# Optional: LLM-as-a-Judge evaluation
# ==========================================================
ragas>=0.2.0
datasets>=2.18.0
langchain-openai>=0.1.0
